requirements.txt
langchain-core
langchain-openai
langchain-community
faiss-cpu
beautifulsoup4
requests
PyMuPDF
pandas
numpy
tiktoken
python-dotenv
tqdm
Flask
Flask-CORS
gunicorn
google-search-results
serpapi
sentence-transformers
